Solve day20 pt.2

+++ src/aoc22/aoc.h
@@ -42,5 +42,5 @@ int day18_1(std::istream *input_file);
 int day18_2(std::istream *input_file);
 int day19_1(std::istream *input_file);
 int day19_2(std::istream *input_file);
-int day20_1(std::istream *input_file);
-int day20_2(std::istream *input_file);
+long day20_1(std::istream *input_file);
+long day20_2(std::istream *input_file);
