+++ src/aoc22/day20.main.cpp
@@ -1,6 +1,6 @@
 #include "aoc22/aoc.h"
 
 int main(int argc, char *argv[]) {
-  run_day<int, int>(argc, argv, day20_1, day20_2);
+  run_day<long, long>(argc, argv, day20_1, day20_2);
   return 0;
 }
